Continue AutoService. Save progress

+++ src/main/java/com/example/shiftplanner_server/services/auto/Score.java
@@ -2,6 +2,6 @@ package com.example.shiftplanner_server.services.auto;
 
 import com.example.shiftplanner_server.entities.Staff;
 
-public record Score(Staff staff, int value) {
+public record Score(Staff staff, long value) {
 }
